changing EmployeeAgeFilter apply button color

+++ lib/features/Employ/repository/employ_repo.dart
@@ -3,5 +3,5 @@ import '../model/employ_model.dart';
 class EmployeeRepo{
   Future<EmployeeData?> getEmployeeData() async {}
 
-  List<Employee>? getFilteredList(double minAge ,double maxAge,List<Employee> epmList){}
+  List<Employee>? getFilteredList(int minAge ,int maxAge,List<Employee> epmList){}
 }
